Added new files for minor refactor
