+++ src/components/ids-data-grid/ids-data-grid.scss
@@ -1,11 +1,12 @@
 @import '../../core/ids-base';
 @import './ids-data-grid-filters';
+@import './ids-data-grid-header';
+@import './ids-data-grid-row';
 @import '../../components/ids-popup-menu/ids-popup-menu';
 @import '../../mixins/sass/ids-checkbox-mixin';
 @import '../../mixins/sass/ids-radio-mixin';
 
 $cell-border-color: var(--ids-color-palette-slate-30);
-$header-bg-color: var(--ids-color-palette-slate-80);
 
 :host {
   --ids-data-grid-column-widths: repeat(1, minmax(110px, 1fr));
@@ -36,291 +37,6 @@ $header-bg-color: var(--ids-color-palette-slate-80);
   @include antialiased();
   @include font-sans();
   @include text-16();
-
-  // Row Styles
-  .ids-data-grid-row {
-    @include w-full();
-
-    display: grid;
-    grid-template-columns: var(--ids-data-grid-column-widths); // Filled in dynamically
-
-    &.selected:not(.mixed) .ids-data-grid-cell,
-    &.activated .ids-data-grid-cell {
-      @include bg-azure-10();
-    }
-
-    &[hidden] {
-      display: none;
-    }
-  }
-
-  // Row Height Feature
-  // Large by default
-  .ids-data-grid-body .ids-data-grid-row .ids-data-grid-cell {
-    height: 50px;
-  }
-
-  &[data-row-height='md'] .ids-data-grid-body .ids-data-grid-row .ids-data-grid-cell {
-    height: 40px;
-  }
-
-  &[data-row-height='sm'] .ids-data-grid-body .ids-data-grid-row .ids-data-grid-cell {
-    height: 35px;
-  }
-
-  &[data-row-height='xs'] .ids-data-grid-body .ids-data-grid-row .ids-data-grid-cell {
-    height: 30px;
-  }
-
-  // Row Alternate Row Setting Feature
-  &.alt-row-shading .ids-data-grid-body [role='row']:nth-child(even) {
-    @include bg-slate-10();
-  }
-
-  // Header Styles
-  .ids-data-grid-header {
-    @include sticky();
-
-    top: 0;
-    z-index: 4; // Higher than the frozen columns
-  }
-
-  // Grouped Headers
-  .ids-data-grid-column-groups {
-    @include border-0();
-    @include border-b-1();
-    @include border-solid();
-
-    border-color: $cell-border-color;
-  }
-
-  // Header Cell Area
-  .ids-data-grid-header-cell {
-    @include border-0();
-    @include border-l-1();
-    @include border-solid();
-
-    background-color: $header-bg-color;
-    border-color: $cell-border-color;
-    min-height: 35px;
-    position: relative;
-
-    // Hide first border
-    &:first-child {
-      border-inline-start-width: 0;
-    }
-
-    &.align-center .ids-data-grid-header-cell-content,
-    &.align-center .ids-data-grid-header-cell-filter-wrapper {
-      justify-content: center;
-    }
-
-    &.align-right .ids-data-grid-header-cell-content,
-    &.align-right .ids-data-grid-header-cell-filter-wrapper {
-      direction: rtl;
-      justify-content: flex-start;
-    }
-
-    // Resizing Cells
-    .resizer {
-      cursor: col-resize;
-      display: inline;
-      height: 100%;
-      position: absolute;
-      margin-inline-start: calc(100% - 4px);
-      top: 0;
-      width: 6px;
-      z-index: 2;
-    }
-
-    &:last-child .resizer {
-      // Prevents un needed scrollbar
-      margin-inline-start: calc(100% - 6px);
-    }
-
-    // Reorderable Cells
-    .reorderer {
-      cursor: move;
-      display: inline;
-      height: 33px;
-      opacity: 0;
-      position: absolute;
-      inset-inline-start: 7px;
-      top: 0;
-      width: 4px;
-
-      ids-icon {
-        position: absolute;
-        color: var(--ids-color-palette-white);
-        inset-inline-start: -7px;
-        text-align: center;
-        top: 9px;
-      }
-    }
-
-    &.align-right .reorderer {
-      inset-inline-end: 7px;
-      inset-inline-start: initial;
-    }
-
-    &.dragging {
-      background-color: $header-bg-color !important;
-      display: block;
-      width: 110px;
-
-      .reorderer {
-        opacity: 1;
-      }
-    }
-
-    &.active-drag-column {
-      background-color: var(--ids-color-palette-slate-40);
-    }
-
-    &:hover .reorderer {
-      opacity: 1;
-    }
-
-    // Reorderable Cells
-    &.frozen {
-      background-color: $header-bg-color;
-      position: sticky;
-      z-index: 3;
-    }
-
-    // Change borders for frozen column scrolling
-    &.frozen-last {
-      border-inline-end-width: 1px;
-
-      + .ids-data-grid-header-cell {
-        border-inline-start-width: 0;
-      }
-    }
-
-    &.frozen-left {
-      position: sticky;
-      left: 0;
-
-      // Dynamically added widths
-      @for $i from 2 through 10 {
-        &:nth-of-type(#{$i}) {
-          left: var(--ids-data-grid-frozen-column-left-width-#{$i});
-        }
-      }
-    }
-
-    &.frozen-right {
-      position: sticky;
-      right: 0;
-
-      // Wierd centering bug (frozen right)
-      .ids-data-grid-header-cell-content {
-        display: inline-flex;
-      }
-    }
-  }
-
-  .ids-data-grid-header-cell-content {
-    @include flex();
-  }
-
-  .ids-data-grid-header-text {
-    @include block();
-    @include px-16();
-    @include py-8();
-    @include font-bold();
-    @include text-14();
-    @include text-white();
-
-    overflow: hidden;
-    text-align: left;
-    text-overflow: ellipsis;
-    user-select: none;
-    white-space: nowrap;
-  }
-
-  .ids-data-grid-header-icon {
-    @include text-white();
-    @include ml-4();
-    @include mt-4();
-
-    align-self: center;
-  }
-
-  // Header sort styles
-  .sort-indicator {
-    @include relative();
-
-    overflow: hidden;
-    height: 32px;
-    margin-inline-start: 2px;
-    width: 18px;
-  }
-
-  // Sort Hover states
-  .is-sortable {
-    .sort-indicator ids-icon {
-      color: var(--ids-color-palette-slate-50);
-      opacity: 0;
-      transition: opacity 0.3s cubic-bezier(0.17, 0.04, 0.03, 0.94) 0s;
-
-      &:first-child {
-        @include relative();
-
-        top: 4px;
-        transform: rotate(180deg);
-      }
-
-      &:last-child {
-        @include relative();
-
-        top: -9px;
-      }
-    }
-
-    &:hover {
-      cursor: pointer;
-
-      .sort-indicator ids-icon {
-        opacity: 1;
-      }
-    }
-  }
-
-  .has-selectioncheckbox .ids-data-grid-header-text {
-    @include px-0();
-  }
-
-  .is-sortable,
-  .has-headericon {
-    .ids-data-grid-header-text {
-      @include pr-0();
-    }
-  }
-
-  // Ascending sort
-  [aria-sort='ascending'] .is-sortable .sort-indicator ids-icon {
-    &:first-child {
-      color: var(--ids-color-palette-white);
-      opacity: 1;
-    }
-
-    &:last-child {
-      opacity: 1;
-    }
-  }
-
-  // Descending sort
-  [aria-sort='descending'] .is-sortable .sort-indicator ids-icon {
-    &:first-child {
-      opacity: 1;
-    }
-
-    &:last-child {
-      color: var(--ids-color-palette-white);
-      opacity: 1;
-    }
-  }
 
   // Cell Styles
   .ids-data-grid-cell {
@@ -633,37 +349,6 @@ $header-bg-color: var(--ids-color-palette-slate-80);
 table.ids-data-grid {
   outline: 1px solid var(--ids-color-palette-slate-30);
   outline-offset: -1px;
-}
-
-// Drag arrows for sorting
-.ids-data-grid-sort-arrows {
-  display: none;
-  height: 34px;
-  position: absolute;
-  left: 25px;
-  z-index: 1;
-
-  &::before,
-  &::after {
-    border-left: 5px solid transparent;
-    border-right: 5px solid transparent;
-    content: '';
-    display: inline-block;
-    height: 0;
-    left: -6px;
-    position: absolute;
-    width: 0;
-  }
-
-  &::before {
-    border-top: 5px solid $header-bg-color;
-    top: -6px;
-  }
-
-  &::after {
-    border-bottom: 5px solid $header-bg-color;
-    bottom: -9px;
-  }
 }
 
 // Dark Theme
